# backend/internal/apidocs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Yivi Business Wallet API
  description: |
    HTTP API for the Yivi Business Wallet backend. All endpoints live under the
    versioned `/api/v1` prefix. Authentication uses the `ybw_session` cookie set
    by the login flow; endpoints that do not require it are marked with no security.
  version: "1.0.0"
servers:
  - url: /api/v1
    description: Versioned API root
...
paths:
...
  /api/v1/orgs/{slug}/identity-proofing/flows/{flowID}/hosted:
    parameters:
      - name: slug
        in: path
        required: true
        schema:
          type: string
        description: Organisation slug.
      - name: flowID
        in: path
        required: true
        schema:
          type: string
        description: The flow's id at the identity-proofing-service.
    get:
      tags: [identity-proofing]
      summary: Read a flow's hosted page settings
      description: >-
        Requires an administrator role in the organisation. `enabled` (hosted links may be made
        for the flow), `locales` (`en`/`nl` the page offers; empty is every one) and
        `completion`: `redirect` (the session's `redirectUrl` when it has one, else the page's
        own thank-you screen) or `done` (always the thank-you screen). A flow never set reads
        `enabled: true, locales: [], completion: redirect`.
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [identity-proofing]
      summary: Save a flow's hosted page settings
      description: >-
        Requires an administrator role in the organisation. Body `{enabled, locales, completion}`
        as read. A hosted session on a flow with its page off is 409 `hosted_disabled`; one with
        a `redirectUrl` on a `done` flow, or a `language` the flow does not offer, is 400.
        Audited `identity_proofing.flow_hosted_configured` with before and after.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: JSON request payload; see the endpoint description.
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/proof/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
        description: Opaque hosted-link token from the `hostedUrl` of a hosted proofing session.
    get:
      tags: [identity-proofing]
      summary: Open a hosted proofing link
      security: []
      description: >-
        The hosted page, unauthenticated: the token is the whole authorization, and every route is rate-limited per link. The customer's `name` and `branding` (its `logoUri` is this link's logo route), what the flow collects (`requestedAttributes`, `requiredAssuranceLevel`, `yiviAvailable`), and the progress: `status`, `started`, and `linkExpiresAt` (the link's end until started, then the session's). `sessionId` is the customer's id for the session; `redirectUrl` (absent for the page's own done screen) where the page sends the subject once settled; `embedOrigins` the customer's allowed redirect origins, the only ones the page posts its `proofing.completed` message to; `language` (`en` or `nl`, absent for the browser's) the page's language, and `locales` the languages the flow's page offers (empty: every one). No personal data.
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/proof/{token}/status:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
        description: Opaque hosted-link token from the `hostedUrl` of a hosted proofing session.
    get:
      tags: [identity-proofing]
      summary: Poll a hosted proofing link
      security: []
      description: >-
        The progress part of `GET /api/v1/proof/{token}`, for the page to poll.
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/proof/{token}/start:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
        description: Opaque hosted-link token from the `hostedUrl` of a hosted proofing session.
    post:
      tags: [identity-proofing]
      summary: Start a hosted proofing link
      security: []
      description: >-
        Body `{method}`: `idem_app` or `yivi_app` (only when the flow's `yiviAvailable`). Creates the session once; an Idem session answers its vcmrtd `deepLink`.
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '409':
          description: The link was started already, its session ended, or the customer is paused.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/proof/{token}/decline:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
        description: Opaque hosted-link token from the `hostedUrl` of a hosted proofing session.
    post:
      tags: [identity-proofing]
      summary: Decline a hosted proofing link
      security: []
      description: >-
        The subject declined what the flow collects: cancels the link before it was started, audited as
        `identity_proofing.session_cancelled` by `hosted_link`. Answers the progress, `status` `cancelled`.
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '409':
          description: The link was started already or its session ended.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/proof/{token}/logo:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
        description: Opaque hosted-link token from the `hostedUrl` of a hosted proofing session.
    get:
      tags: [identity-proofing]
      summary: A hosted link's customer logo
      security: []
      description: >-
        The customer's logo image, for the hosted page.
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/proof/{token}/claim-link:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
        description: Opaque hosted-link token from the `hostedUrl` of a hosted proofing session.
    post:
      tags: [identity-proofing]
      summary: Get a fresh Idem app link for a hosted link
      security: []
      description: >-
        As the org route of the same name, for a hosted link started with `idem_app`.
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '409':
          description: The link does not run in the Idem app (`wrong_method`), its session has ended (`session_over`), or the app holding it is still active (`device_active`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/proof/{token}/yivi/start:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
        description: Opaque hosted-link token from the `hostedUrl` of a hosted proofing session.
    post:
      tags: [identity-proofing]
      summary: Start the Yivi disclosure of a hosted link
      security: []
      description: >-
        As the org route of the same name, for a hosted link started with `yivi_app`.
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '409':
          description: The link was started already, its session ended, or the customer is paused.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/proof/{token}/yivi/disclosure:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
        description: Opaque hosted-link token from the `hostedUrl` of a hosted proofing session.
    get:
      tags: [identity-proofing]
      summary: Poll the Yivi disclosure of a hosted link
      security: []
      description: >-
        As the org route of the same name, for a hosted link.
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '409':
          description: The link was started already, its session ended, or the customer is paused.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/proof/{token}/yivi/face:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
        description: Opaque hosted-link token from the `hostedUrl` of a hosted proofing session.
    post:
      tags: [identity-proofing]
      summary: Send a face frame of a hosted link
      security: []
      description: >-
        As the org route of the same name, for a hosted link: body `{image}`, one camera frame.
      responses:
        '2XX':
          description: Successful response.
          content:
            application/json:
              schema:
                description: Response payload; shape depends on the endpoint.
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '409':
          description: The link was started already, its session ended, or the customer is paused.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    customerApiKey:
      type: http
      scheme: bearer
      description: A proofing customer's API key (`yp_live_…`), for the public `/api/v1/proofing` routes.
    cookieAuth:
      type: apiKey
      in: cookie
      name: ybw_session
  schemas:
    Error:
      type: object
      description: Error response body returned by respond.Error.
      properties:
        error:
          type: string
          description: Human-readable error message.
        code:
          type: string
          description: Stable machine-readable error code.
      required: [error, code]
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: No valid session cookie was supplied.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: The session lacks the required role for this resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The referenced resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: >-
        The customer spent its public API budget (code `rate_limited`): 120 calls a
        minute, and of those 10 session creations. `Retry-After` says in how many
        seconds to try again.
      headers:
        Retry-After:
          description: Seconds until the next call is allowed.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
security:
  - cookieAuth: []